ICU-13725 Additional MSVC compatibility fixes.
X-SVN-Rev: 41272

+++ icu4c/source/i18n/numparse_parsednumber.cpp
@@ -72,7 +72,7 @@ double ParsedNumber::getDouble() const {
     }
 
     if (quantity.fitsInLong()) {
-        return quantity.toLong();
+        return static_cast<double>(quantity.toLong());
     } else {
         return quantity.toDouble();
     }
